+++ bin/tsg_pagerank
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+
+import logging
+import sys
+
+import pandas as pd
+
+from tsg.intermediate import generate_intermediate
+from tsg.config import RAW_DIR, PAGERANK_PATH
+from tsg.indexer.page_rank import # specify
+
+logging.basicConfig(level=logging.INFO)
+
+
+def main():
+    print('Do you want to delete pagerank file? y/N')
+    if input() != 'y':
+        sys.exit(0)
+
+    # get page rank
+    page_rank_scores = call_your_function
+
+    # convert to dataframe
+    df = pd.DataFrame()
+
+    df.to_csv(PAGERANK_PATH)
+
+
+if __name__ == "__main__":
+    main()
