+++ tsg/config.py
@@ -25,9 +25,9 @@ FIELD_WEIGHTS = [2.5, 2.5, 5, 1]
 CSV_HEADER = ','.join(['uuid'] + FIELDS)
 
 DICTIONARY_PATH = DATA_DIR + 'dictionary.dat'
+PAGERANK_PATH = DATA_DIR + 'pagerank.json'
 INDEXINFO_PATH = DATA_DIR + 'indexinfo.json'
 
-THROTTLE_SECONDS = 1.6  # default throttle time, gets replaced by robots.txt,
-                        # parsing
+THROTTLE_SECONDS = 1.6  # default throttle time, gets replaced by robots.txt
 ALLOWED_SITES = []
 DISALLOWED_SITES = []
